fix(source): Fixed the parse binary timestamp bug: added parse integer step for the seconds portion
The existing implementation parses the seconds portion of the timestamp as a whole decimal number,
whereas the seconds portion is actually encoded as an integer portion plus an fractional portion.
Therefore, the seconds portion should be first parsed as an integer, and if there is additional data
in the span (i.e. the fraction portion), the remaining data should then be parsed as a decimal
number same as the existing implementation.