Fixes decimal tests to work across Node 8/10/12. (#409)
The underlying issue is that we test using `Decimal.numberValue()`
which uses `Math.pow(10, exponent)`.  This API has different rounding
behaviors (which is allowed by the ECMAScript spec) in v8/v10 of
Node vs. v12 of Node.  This is likely to be an issue on different
engines as well.

The fix allows the `test` method to support an inclusive interval of
valid `number` values that the `numberValue()` representation can be
and fixes the one test that seems to trip on this.

Resolves #399