Remove dependency on `global` global variable (#715)
The `global` global variable [1] is a node-ism, and is not available on
browsers and other js runtime environments. This issue is typically
overcome in browsers with a polyfill that defines `global` as a
reference to the `window` variable [2], but not all runtimes have an
equivalent option. As a result, use of `global` tends to cause headaches:

https://github.com/aws/aws-sdk-js/issues/2141

https://github.com/angular/angular-cli/issues/8160

https://stackoverflow.com/questions/57361546/referenceerror-global-is-not-defined-with-web3

https://dev.to/richardbray/how-to-fix-the-referenceerror-global-is-not-defined-error-in-sveltekitvite-2i49

Therefore, this change replaces all references to `global` with
behavior that is guaranteed by the ECMA spec, which should be available
in any runtime.

[1]: https://nodejs.org/docs/latest/api/all.html#all_globals_global
[2]: https://github.com/browserify/browserify-handbook#global